+++ CHANGELOG.md
@@ -28,13 +28,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Tooltips to support usability when datalist is not supported
 - Added options for Test and QA buckets as targets for Jenkins builds
 - Added support for Ground Water Site Attributes
-- Changed styling so all elements are left justified
+- Added support for National and Local Aquifer codes
 - Added site status parameter
 - Added watershed area query
 - Added Altitude query params
 - Added csv processing for quickly adding a lot codes in fields which accept and verify multiple codes
 - Added various temporal query parameters
-### changed
+### Changed
+- Changed styling so all elements are left justified
 - separted sitelist from main.vue
 - switched datetime library to momentjs from date-format
 - separated sitelist from main.vue